// src/types/mapbox-gl.d.ts
    export type ExpressionName =
        // Types
        | 'array' | 'boolean' | 'collator' | 'format' | 'literal' | 'number' | 'object' | 'string'
        | 'to-boolean' | 'to-color' | 'to-number' | 'to-string' | 'typeof'
        // Feature data
        | 'feature-state' | 'geometry-type' | 'id' | 'line-progress' | 'properties'
        // Lookup
        | 'at' | 'get' | 'has' | 'length'
        // Decision
        | '!' | '!=' | '<' | '<=' | '==' | '>' | '>=' | 'all' | 'any' | 'case' | 'match' | 'coalesce'
        // Ramps, scales, curves
        | 'interpolate' | 'interpolate-hcl' | 'interpolate-lab' | 'step'
        // Variable binding
        | 'let' | 'var'
        // String
        | 'concat' | 'downcase' | 'is-supported-script' | 'resolved-locale' | 'upcase'
        // Color
        | 'rgb' | 'rgba'
        // Math
        | '-' | '*' | '/' | '%' | '^' | '+' | 'abs' | 'acos' | 'asin' | 'atan' | 'ceil' | 'cos' | 'e'
        | 'floor' | 'ln' | 'ln2' | 'log10' | 'log2' | 'max' | 'min' | 'pi' | 'round' | 'sin' | 'sqrt' | 'tan'
        // Zoom, Heatmap
        | 'zoom' | 'heatmap-density';

    export type Expression = [ExpressionName, ...any[]]

    type Anchor = 'center' | 'left' | 'right' | 'top' | 'bottom' | 'top-left' | 'top-right' | 'bottom-left' | 'bottom-right';
// ...
    export interface StyleFunction {
        stops?: any[][];
        property?: string;
        base?: number;
        type?: 'identity' | 'exponential' | 'interval' | 'categorical';
        default?: any;
        'colorSpace'?: 'rgb' | 'lab' | 'hcl';
    }

    type Visibility = 'visible' | 'none';

    export interface Layout {
        visibility?: Visibility;
    }
// ...
    export interface SymbolLayout extends Layout {
        'symbol-placement'?: 'point' | 'line' | 'line-center';
        'symbol-spacing'?: number | Expression;
        'symbol-avoid-edges'?: boolean;
        'symbol-z-order'?: 'viewport-y' | 'source';
        'icon-allow-overlap'?: boolean | StyleFunction | Expression;
        'icon-ignore-placement'?: boolean;
        'icon-optional'?: boolean;
        'icon-rotation-alignment'?: 'map' | 'viewport' | 'auto';
        'icon-size'?: number | StyleFunction | Expression;
        'icon-text-fit'?: 'none' | 'both' | 'width' | 'height';
        'icon-text-fit-padding'?: number[] | Expression;
        'icon-image'?: string | StyleFunction | Expression;
        'icon-rotate'?: number | StyleFunction | Expression;
        'icon-padding'?: number | Expression;
        'icon-keep-upright'?: boolean;
        'icon-offset'?: number[] | StyleFunction | Expression;
        'icon-anchor'?: Anchor | StyleFunction | Expression;
        'icon-pitch-alignment'?: 'map' | 'viewport' | 'auto';
        'text-pitch-alignment'?: 'map' | 'viewport' | 'auto';
        'text-rotation-alignment'?: 'map' | 'viewport' | 'auto';
        'text-field'?: string | StyleFunction | Expression;
        'text-font'?: string | string[] | Expression;
        'text-size'?: number | StyleFunction | Expression;
        'text-max-width'?: number | Expression;
        'text-line-height'?: number | Expression;
        'text-letter-spacing'?: number | Expression;
        'text-justify'?: 'left' | 'center' | 'right' | Expression;
        'text-anchor'?: Anchor | StyleFunction | Expression;
        'text-max-angle'?: number | Expression;
        'text-rotate'?: number | StyleFunction | Expression;
        'text-padding'?: number | Expression;
        'text-keep-upright'?: boolean;
        'text-transform'?: 'none' | 'uppercase' | 'lowercase' | StyleFunction | Expression;
        'text-offset'?: number[] | Expression;
        'text-allow-overlap'?: boolean;
        'text-ignore-placement'?: boolean;
        'text-optional'?: boolean;
    }